Updating the code to incorperate the new changes made in the mapper and the assetquote files

+++ backend/fortunelink/src/main/java/com/laderrco/fortunelink/portfolio_management/infrastructure/external/marketdata/fmp/dtos/FmpQuoteResponse.java
@@ -70,7 +70,7 @@ public class FmpQuoteResponse {
     private BigDecimal yearLow;
 
     @JsonProperty("marketCap")
-    private Long marketCap;
+    private BigDecimal marketCap;
 
     @JsonProperty("priceAvg50")
     private BigDecimal priceAvg50;
